move add to trial session modal values into separate computed; group by state in dropdown
