+++ web-client/src/presenter/computeds/caseDetailHelper.test.js
@@ -703,182 +703,6 @@ describe('case detail computed', () => {
     expect(result.respondentSearchResultsCount).toBeUndefined();
   });
 
-  describe('trialSessionsFormatted', () => {
-    const trialSessions = [
-      {
-        sessionType: 'Regular',
-        startDate: '2019-03-01T21:40:46.415Z',
-        status: 'Upcoming',
-        trialLocation: 'Birmingham, Alabama',
-        trialSessionId: '1',
-      },
-      {
-        sessionType: 'Hybrid',
-        startDate: '2018-02-01T21:40:46.415Z',
-        status: 'Upcoming',
-        trialLocation: 'Mobile, Alabama',
-        trialSessionId: '2',
-      },
-      {
-        sessionType: 'Special',
-        startDate: '2019-01-01T21:40:46.415Z',
-        status: 'Upcoming',
-        trialLocation: 'Birmingham, Alabama',
-        trialSessionId: '3',
-      },
-      {
-        sessionType: 'Small',
-        startDate: '2019-05-01T21:40:46.415Z',
-        status: 'Upcoming',
-        trialLocation: 'Boise, Idaho',
-        trialSessionId: '4',
-      },
-      {
-        sessionType: 'Motion/Hearing',
-        startDate: '2018-12-01T21:40:46.415Z',
-        status: 'Upcoming',
-        trialLocation: 'Mobile, Alabama',
-        trialSessionId: '5',
-      },
-    ];
-
-    it('should not return trialSessionsFormatted if modal state does not contain trialSessions', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: {},
-          form: {},
-          modal: {},
-        },
-      });
-
-      expect(result.trialSessionsFormatted).toBeFalsy();
-    });
-
-    it('should filter out trial sessions that do not have the status Upcoming', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: { preferredTrialCity: 'Birmingham, Alabama' },
-          form: {},
-          modal: {
-            showAllLocations: true,
-            trialSessions: [
-              ...trialSessions,
-              {
-                status: 'Closed',
-                trialLocation: 'Nashville, Tennessee',
-                trialSessionId: '5',
-              },
-            ],
-          },
-        },
-      });
-
-      expect(result.trialSessionsFormatted.length).toEqual(
-        trialSessions.length,
-      );
-    });
-
-    it('should filter trialSessions by preferredTrialCity if state.modal.showAllLocations is false', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: { preferredTrialCity: 'Birmingham, Alabama' },
-          form: {},
-          modal: {
-            showAllLocations: false,
-            trialSessions,
-          },
-        },
-      });
-
-      expect(result.trialSessionsFormatted.length).toEqual(2);
-      expect(result.trialSessionsFormatted).toMatchObject([
-        { trialLocation: 'Birmingham, Alabama' },
-        { trialLocation: 'Birmingham, Alabama' },
-      ]);
-    });
-
-    it('should not filter trialSessions by preferredTrialCity if state.modal.showAllLocations is true', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: { preferredTrialCity: 'Birmingham, Alabama' },
-          form: {},
-          modal: {
-            showAllLocations: true,
-            trialSessions,
-          },
-        },
-      });
-
-      expect(result.trialSessionsFormatted.length).toEqual(
-        trialSessions.length,
-      );
-    });
-
-    it('should format optionText for each trial session and sort by location and then by date when showAllLocations is true', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: { preferredTrialCity: 'Birmingham, Alabama' },
-          form: {},
-          modal: {
-            showAllLocations: true,
-            trialSessions,
-          },
-        },
-      });
-
-      expect(result.trialSessionsFormatted.length).toEqual(
-        trialSessions.length,
-      );
-      expect(result.trialSessionsFormatted).toMatchObject([
-        {
-          optionText: 'Birmingham, Alabama 01/01/19 (SP)',
-          trialSessionId: '3',
-        },
-        {
-          optionText: 'Birmingham, Alabama 03/01/19 (R)',
-          trialSessionId: '1',
-        },
-        {
-          optionText: 'Boise, Idaho 05/01/19 (S)',
-          trialSessionId: '4',
-        },
-        {
-          optionText: 'Mobile, Alabama 02/01/18 (H)',
-          trialSessionId: '2',
-        },
-        {
-          optionText: 'Mobile, Alabama 12/01/18 (M/H)',
-          trialSessionId: '5',
-        },
-      ]);
-    });
-
-    it('should format optionText for each trial session and sort by date when showAllLocations is false', () => {
-      const result = runCompute(caseDetailHelper, {
-        state: {
-          caseDetail: { preferredTrialCity: 'Birmingham, Alabama' },
-          form: {},
-          modal: {
-            showAllLocations: false,
-            trialSessions,
-          },
-        },
-      });
-
-      expect(result.trialSessionsFormatted.length).toEqual(2);
-      expect(result.trialSessionsFormatted).toMatchObject([
-        {
-          optionText: 'Birmingham, Alabama 01/01/19 (SP)',
-          trialSessionId: '3',
-        },
-        {
-          optionText: 'Birmingham, Alabama 03/01/19 (R)',
-          trialSessionId: '1',
-        },
-      ]);
-    });
-  });
-
   describe('hasOrders', () => {
     it('should signify a given case has orders (hasOrders=TRUE) if any of the order-related props are true', () => {
       const result = runCompute(caseDetailHelper, {
