+++ web-client/src/presenter/state.js
@@ -4,6 +4,7 @@ import {
 } from './computeds/formattedCaseDetail';
 
 import { addDocketEntryHelper } from './computeds/addDocketEntryHelper';
+import { addToTrialSessionModalHelper } from './computeds/addToTrialSessionModalHelper';
 import { advancedSearchHelper } from './computeds/advancedSearchHelper';
 import { alertHelper } from './computeds/alertHelper';
 import { blockedCasesReportHelper } from './computeds/blockedCasesReportHelper';
@@ -52,6 +53,7 @@ import { workQueueSectionHelper } from './computeds/workQueueSectionHelper';
 
 export const state = {
   addDocketEntryHelper,
+  addToTrialSessionModalHelper,
   advancedSearchForm: {},
   advancedSearchHelper,
   alertHelper,
